Added more bug fixes.
Tweaking a bit more. Adding mysql database and configs for wp. This may not work by running as root - the sudo command with a function is a guess. It kinda runs from the command line - but if this does not work, I can at least remember the commands I ran manually.